requirements.txt
pytest
requests
flask
ipywidgets
PyYAML
pandas
plotly
